@@ -29,8 +29,8 @@
     "symfony/event-dispatcher": "^7.3",
     "symfony/event-dispatcher-contracts": "^3",
     "symfony/form": "^7.3",
-    "symfony/framework-bundle": "^7.3",
-    "symfony/http-foundation": "^7.3",
+    "symfony/framework-bundle": "^7.4",
+    "symfony/http-foundation": "^7.4",
     "symfony/http-kernel": "^7.3",
     "symfony/lock": "^7.3",
     "symfony/messenger": "^7.3",
@@ -56,7 +56,7 @@
     "tourze/easy-admin-menu-bundle": "1.0.*",
     "tourze/enum-extra": "1.0.*",
     "tourze/json-rpc-cache-bundle": "1.0.*",
-    "tourze/json-rpc-core": "1.0.*",
+    "tourze/json-rpc-core": "2.0.*",
     "tourze/json-rpc-lock-bundle": "1.0.*",
     "tourze/json-rpc-log-bundle": "1.0.*",
     "tourze/json-rpc-paginator-bundle": "1.0.*",
@@ -76,9 +76,10 @@
     "tourze/phpunit-base": "1.*",
     "tourze/phpunit-doctrine-entity": "1.*",
     "tourze/phpunit-enum": "1.*",
-    "tourze/phpunit-symfony-kernel-test": "1.0.*",
-    "tourze/phpunit-symfony-unit-test": "1.*",
-    "tourze/phpunit-symfony-web-test": "1.*"
+    "tourze/phpunit-json-rpc": "1.0.*",
+    "tourze/phpunit-symfony-kernel-test": "2.0.*",
+    "tourze/phpunit-symfony-unit-test": "2.0.*",
+    "tourze/phpunit-symfony-web-test": "2.0.*"
   },
   "suggest": {
     "tourze/cms-collect-bundle": "为CMS内容提供收藏功能",
@@ -86,12 +87,12 @@
   },
   "autoload": {
     "psr-4": {
-      "CmsBundle\\": "src"
+      "Tourze\\CmsBundle\\": "src"
     }
   },
   "autoload-dev": {
     "psr-4": {
-      "CmsBundle\\Tests\\": "tests"
+      "Tourze\\CmsBundle\\Tests\\": "tests"
     }
   }
 }
