@@ -2,12 +2,12 @@
 
 declare(strict_types=1);
 
-namespace CmsBundle\Exception;
+namespace Tourze\CmsBundle\Exception;
 
 /**
  * 当CMS模型未找到时抛出的异常.
  */
-class ModelNotFoundException extends \RuntimeException
+final class ModelNotFoundException extends \RuntimeException
 {
     public function __construct(string $modelCode, ?\Throwable $previous = null)
     {
