@@ -2,11 +2,11 @@
 
 declare(strict_types=1);
 
-namespace CmsBundle\Tests;
+namespace Tourze\CmsBundle\Tests;
 
-use CmsBundle\CmsBundle;
 use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\Attributes\RunTestsInSeparateProcesses;
+use Tourze\CmsBundle\CmsBundle;
 use Tourze\PHPUnitSymfonyKernelTest\AbstractBundleTestCase;
 
 /**
