@@ -2,17 +2,17 @@
 
 declare(strict_types=1);
 
-namespace CmsBundle\Tests\Controller\Admin;
+namespace Tourze\CmsBundle\Tests\Controller\Admin;
 
-use CmsBundle\Controller\Admin\AttributeCrudController;
-use CmsBundle\Entity\Attribute;
-use CmsBundle\Entity\Model;
-use CmsBundle\Enum\FieldType;
-use CmsBundle\Tests\AbstractCmsControllerTestCase;
 use EasyCorp\Bundle\EasyAdminBundle\Controller\AbstractCrudController;
 use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\Attributes\RunTestsInSeparateProcesses;
 use Symfony\Component\Security\Core\Exception\AccessDeniedException;
+use Tourze\CmsBundle\Controller\Admin\AttributeCrudController;
+use Tourze\CmsBundle\Entity\Attribute;
+use Tourze\CmsBundle\Entity\Model;
+use Tourze\CmsBundle\Enum\FieldType;
+use Tourze\CmsBundle\Tests\AbstractCmsControllerTestCase;
 
 /**
  * @internal
@@ -45,7 +45,7 @@ final class AttributeCrudControllerTest extends AbstractCmsControllerTestCase
 
     public function testCrudConfiguration(): void
     {
-        $controller = new AttributeCrudController();
+        $controller = self::getService(AttributeCrudController::class);
         $this->assertInstanceOf(AttributeCrudController::class, $controller);
     }
 
@@ -386,13 +386,8 @@ final class AttributeCrudControllerTest extends AbstractCmsControllerTestCase
             return $this->cachedController;
         }
 
-        // 确保创建客户端（会自动设置为全局客户端）
-        $client = static::createClient();
-        $container = $client->getContainer();
-        $adminUrlGenerator = $container->get(\EasyCorp\Bundle\EasyAdminBundle\Router\AdminUrlGenerator::class);
-        $this->assertInstanceOf(\EasyCorp\Bundle\EasyAdminBundle\Router\AdminUrlGenerator::class, $adminUrlGenerator);
-
-        $this->cachedController = new AttributeCrudController();
+        // 从容器中获取服务，而不是直接实例化
+        $this->cachedController = self::getService(AttributeCrudController::class);
 
         return $this->cachedController;
     }
