@@ -2,15 +2,15 @@
 
 declare(strict_types=1);
 
-namespace CmsBundle\Tests\Controller\Admin;
+namespace Tourze\CmsBundle\Tests\Controller\Admin;
 
-use CmsBundle\Controller\Admin\VisitStatCrudController;
-use CmsBundle\Entity\VisitStat;
-use CmsBundle\Tests\AbstractCmsControllerTestCase;
 use EasyCorp\Bundle\EasyAdminBundle\Controller\AbstractCrudController;
 use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\Attributes\RunTestsInSeparateProcesses;
 use Symfony\Component\Security\Core\Exception\AccessDeniedException;
+use Tourze\CmsBundle\Controller\Admin\VisitStatCrudController;
+use Tourze\CmsBundle\Entity\VisitStat;
+use Tourze\CmsBundle\Tests\AbstractCmsControllerTestCase;
 
 /**
  * @internal
@@ -43,7 +43,7 @@ final class VisitStatCrudControllerTest extends AbstractCmsControllerTestCase
 
     public function testCrudConfiguration(): void
     {
-        $controller = new VisitStatCrudController();
+        $controller = self::getService(VisitStatCrudController::class);
         $this->assertInstanceOf(VisitStatCrudController::class, $controller);
     }
 
@@ -92,13 +92,8 @@ final class VisitStatCrudControllerTest extends AbstractCmsControllerTestCase
             return $this->cachedController;
         }
 
-        // 确保创建客户端（会自动设置为全局客户端）
-        $client = static::createClient();
-        $container = $client->getContainer();
-        $adminUrlGenerator = $container->get(\EasyCorp\Bundle\EasyAdminBundle\Router\AdminUrlGenerator::class);
-        $this->assertInstanceOf(\EasyCorp\Bundle\EasyAdminBundle\Router\AdminUrlGenerator::class, $adminUrlGenerator);
-
-        $this->cachedController = new VisitStatCrudController();
+        // 从容器中获取服务，而不是直接实例化
+        $this->cachedController = self::getService(VisitStatCrudController::class);
 
         return $this->cachedController;
     }
