@@ -2,16 +2,16 @@
 
 declare(strict_types=1);
 
-namespace CmsBundle\Tests\Repository;
+namespace Tourze\CmsBundle\Tests\Repository;
 
-use CmsBundle\Entity\Entity;
-use CmsBundle\Entity\Model;
-use CmsBundle\Enum\EntityState;
-use CmsBundle\Repository\EntityRepository;
-use CmsBundle\Repository\ModelRepository;
 use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
 use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\Attributes\RunTestsInSeparateProcesses;
+use Tourze\CmsBundle\Entity\Entity;
+use Tourze\CmsBundle\Entity\Model;
+use Tourze\CmsBundle\Enum\EntityState;
+use Tourze\CmsBundle\Repository\EntityRepository;
+use Tourze\CmsBundle\Repository\ModelRepository;
 use Tourze\PHPUnitSymfonyKernelTest\AbstractRepositoryTestCase;
 
 /**
